code formatted

+++ cloudreve-install.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-cat << EOF
+cat <<EOF
 #
 # cloudreve-install.sh
 # Support OS: Debian / Ubuntu / CentOS
@@ -10,11 +10,11 @@ cat << EOF
 EOF
 
 no_command() {
-    if ! command -v $1 > /dev/null 2>&1; then
+    if ! command -v $1 >/dev/null 2>&1; then
         if [ -z "$3" ]; then
-        $2 install -y $1
+            $2 install -y $1
         else
-        $2 install -y $3
+            $2 install -y $3
         fi
     fi
 }
@@ -22,58 +22,56 @@ no_command() {
 read -p "Please press \"y\" to continue: " answer
 
 case $answer in
-    Y|y)
+Y | y)
     echo "continue..."
 
-#check OS
-source /etc/os-release
+    #check OS
+    source /etc/os-release
 
-        case $ID in
-        debian|ubuntu|devuan)
+    case $ID in
+    debian | ubuntu | devuan)
         echo System OS is $PRETTY_NAME
-    apt update
-    no_command wget apt
-    no_command curl apt
-    no_command tar apt
-    no_command pkill apt procps
+        apt update
+        no_command wget apt
+        no_command curl apt
+        no_command tar apt
+        no_command pkill apt procps
         ;;
 
-        centos|fedora|rhel|sangoma)
+    centos | fedora | rhel | sangoma)
         echo System OS is $PRETTY_NAME
-    no_command bc yum
-    yumdnf="yum"
-    if test "$(echo "$VERSION_ID >= 22" | bc)" -ne 0; then
-        yumdnf="dnf"
-    fi
-    no_command wget $yumdnf
-    no_command curl $yumdnf
-    no_command tar $yumdnf
-    no_command pkill $yumdnf procps-ng
+        no_command bc yum
+        yumdnf="yum"
+        if test "$(echo "$VERSION_ID >= 22" | bc)" -ne 0; then
+            yumdnf="dnf"
+        fi
+        no_command wget $yumdnf
+        no_command curl $yumdnf
+        no_command tar $yumdnf
+        no_command pkill $yumdnf procps-ng
         ;;
-        esac
+    esac
 
-cd ~
-curl -s https://api.github.com/repos/cloudreve/Cloudreve/releases/latest \
-  | grep browser_download_url \
-  | grep linux_amd64 \
-  | cut -d '"' -f 4 \
-  | wget -O cloudreve_linux_amd64.tar.gz -qi - 
+    cd ~
+    curl -s https://api.github.com/repos/cloudreve/Cloudreve/releases/latest |
+        grep browser_download_url |
+        grep linux_amd64 |
+        cut -d '"' -f 4 |
+        wget -O cloudreve_linux_amd64.tar.gz -qi -
 
+    mkdir -p /var/www/cloudreve
+    tar -zxvf cloudreve_linux_amd64.tar.gz -C /var/www/cloudreve
+    chown -R www-data:www-data /var/www/cloudreve
+    chmod 750 /var/www/cloudreve/cloudreve
 
-mkdir -p /var/www/cloudreve
-tar -zxvf cloudreve_linux_amd64.tar.gz -C /var/www/cloudreve
-chown -R www-data:www-data /var/www/cloudreve
-chmod 750 /var/www/cloudreve/cloudreve
+    echo "Please wait Cloudreve init..."
+    runuser -u www-data nohup /var/www/cloudreve/cloudreve >/var/www/cloudreve/cloudreve-install-info.txt 2>&1 &
+    sleep 5s
+    pkill -f /var/www/cloudreve/cloudreve
+    sleep 5s
 
-echo "Please wait Cloudreve init..."
-runuser -u www-data nohup /var/www/cloudreve/cloudreve > /var/www/cloudreve/cloudreve-install-info.txt 2>&1 &
-sleep 5s
-pkill -f /var/www/cloudreve/cloudreve
-sleep 5s
-
-
-# create systemd file and auto run
-cat > /etc/systemd/system/cloudreve.service << EOF
+    # create systemd file and auto run
+    cat >/etc/systemd/system/cloudreve.service <<EOF
 [Unit]
 Description=Cloudreve
 Documentation=https://docs.cloudreve.org
@@ -96,15 +94,14 @@ StandardError=syslog
 WantedBy=multi-user.target
 EOF
 
+    systemctl daemon-reload
+    systemctl enable cloudreve
+    systemctl restart cloudreve
+    systemctl status cloudreve --no-pager
 
-systemctl daemon-reload
-systemctl enable cloudreve
-systemctl restart cloudreve
-systemctl status cloudreve --no-pager
+    cat /var/www/cloudreve/cloudreve-install-info.txt
 
-cat /var/www/cloudreve/cloudreve-install-info.txt
-
-cat << EOF
+    cat <<EOF
 
 =======================================================================
 Cloudreve path        : /var/www/cloudreve/cloudreve
@@ -115,14 +112,11 @@ Default password      : /var/www/cloudreve/cloudreve-install-info.txt
 
 EOF
 
-
-
-## go exit
+    ## go exit
     ;;
 
-
-## end    
-    *)
+    ## end
+*)
     echo "exit"
     ;;
